+++ core/splice-provider/src/index.ts
@@ -2,8 +2,6 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { SpliceProvider } from './SpliceProvider'
-import { SpliceProviderHttp } from './SpliceProviderHttp'
-import { SpliceProviderWindow } from './SpliceProviderWindow'
 
 declare global {
     interface Window {
@@ -18,31 +16,13 @@ export enum ProviderType {
     HTTP,
 }
 
-export function injectSpliceProvider(
-    providerType: ProviderType,
-    url?: URL,
-    sessionToken?: string
-): SpliceProvider {
+export function injectSpliceProvider(provider: SpliceProvider): SpliceProvider {
     // Check if the provider is already injected
     const existing = window.splice || window.canton || window.ethereum
     if (existing) return existing
 
     // Inject the SpliceProvider instance
-    if (providerType === ProviderType.WINDOW) {
-        window.splice = new SpliceProviderWindow()
-    } else if (providerType === ProviderType.HTTP) {
-        if (!url || !(url instanceof URL)) {
-            throw new Error(
-                'Invalid URL for HTTP provider. Please provide a valid URL instance.'
-            )
-        }
-        window.splice = new SpliceProviderHttp(url, sessionToken)
-    } else {
-        throw new Error(
-            'Invalid provider type. Use ProviderType.WINDOW or ProviderType.HTTP.'
-        )
-    }
-
+    window.splice = provider
     window.canton = window.splice // For compatibility with Canton dApps
     window.ethereum = window.splice // For EIP-1193 compatibility
